#430 - move downsampling to data service
