+++ scylla-server/src/services/mod.rs
@@ -1,4 +1,3 @@
 pub mod data_service;
 pub mod data_type_service;
-pub mod downsampling_service;
 pub mod run_service;
