+++ scylla-server/tests/data_service_test.rs
@@ -4,7 +4,7 @@ mod test_utils;
 use scylla_server::{
     error::ScyllaError,
     models::Data,
-    services::{data_service, data_type_service, downsampling_service, run_service},
+    services::{data_service, data_type_service, run_service},
     transformers::data_transformer::{DownsamplingInfo, PublicData},
     ClientData,
 };
@@ -132,22 +132,22 @@ async fn test_data_no_prereqs() -> Result<(), diesel::result::Error> {
 async fn test_downsampling_sampling_rate_calculation() {
     // Test core sampling rate calculation logic
     assert_eq!(
-        downsampling_service::calculate_auto_sampling_rate(5000),
+        data_service::calculate_auto_sampling_rate(5000),
         1,
         "Small dataset should not be downsampled"
     );
     assert_eq!(
-        downsampling_service::calculate_auto_sampling_rate(10000),
+        data_service::calculate_auto_sampling_rate(10000),
         1,
         "At threshold should not be downsampled"
     );
     assert_eq!(
-        downsampling_service::calculate_auto_sampling_rate(15000),
+        data_service::calculate_auto_sampling_rate(15000),
         3,
         "15k points should have rate 3"
     );
     assert_eq!(
-        downsampling_service::calculate_auto_sampling_rate(50000),
+        data_service::calculate_auto_sampling_rate(50000),
         10,
         "50k points should have rate 10"
     );
